reafactor: create use imges hook
