+++ src/public/hooks/useFakeData.ts
@@ -1,22 +1,6 @@
 import { useEffect, useState } from 'react';
 import fakeService from '../sevices/fake-service';
-
-interface Urls {
-  raw: string;
-  full: string;
-  regular: string;
-  small: string;
-  small_s3: string;
-  thumb: string;
-}
-
-export interface UnsplashData {
-  id: string;
-  created_at: string;
-  description: string;
-  alt_description: string;
-  urls: Urls;
-}
+import { UnsplashData } from './useImages';
 
 export const useFakeData = () => {
   const [data, setData] = useState<UnsplashData[]>([]);
